+++ CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Added `TemporalAccessor` support to `SafeString`.
 - Added log4j2 `Marker` support.
+- Added "Copy logger name" accelerator "command shift N".
+- Added "Copy message" accelerator "command shift C".
+- Added "Copy message pattern" accelerator "command shift alt C".
 
 ### Changed
 - Using `java.time.format` instead of `SimpleDateFormat`.